backend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
amqp=5.2.0=pypi_0
annotated-types=0.6.0=pypi_0
anyio=4.2.0=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
billiard=4.2.0=pypi_0
blas=1.0=mkl
boto3=1.34.35=pypi_0
botocore=1.34.35=pypi_0
brotli-python=1.0.9=py310hd77b12b_7
bzip2=1.0.8=he774522_0
ca-certificates=2023.12.12=haa95532_0
celery=5.3.6=pypi_0
certifi=2023.11.17=py310haa95532_0
cffi=1.16.0=py310h2bbff1b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
click-didyoumean=0.3.0=pypi_0
click-plugins=1.1.1=pypi_0
click-repl=0.3.0=pypi_0
clip=1.0=pypi_0
colorama=0.4.6=pypi_0
cpuonly=2.0=0
cryptography=41.0.7=py310h89fc84f_0
environs=9.5.0=pypi_0
exceptiongroup=1.2.0=pypi_0
fastapi=0.109.2=pypi_0
filelock=3.13.1=py310haa95532_0
freetype=2.12.1=ha860e81_0
fsspec=2024.2.0=pypi_0
ftfy=6.1.3=pypi_0
giflib=5.2.1=h8cc25b3_3
gmpy2=2.1.2=py310h7f96b67_0
grpcio=1.60.0=pypi_0
h11=0.14.0=pypi_0
idna=3.4=py310haa95532_0
intel-openmp=2023.1.0=h59b6b97_46320
jinja2=3.1.3=py310haa95532_0
jmespath=1.0.1=pypi_0
jpeg=9e=h2bbff1b_1
kombu=5.3.5=pypi_0
lerc=3.0=hd77b12b_0
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_0
libjpeg-turbo=2.0.0=h196d8e1_0
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
markupsafe=2.1.3=py310h2bbff1b_0
marshmallow=3.20.2=pypi_0
minio=7.2.3=pypi_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py310h2bbff1b_1
mkl_fft=1.3.8=py310h2bbff1b_0
mkl_random=1.2.4=py310h59b6b97_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpir=3.0.0=hec2e145_1
mpmath=1.3.0=py310haa95532_0
networkx=3.1=py310haa95532_0
numpy=1.26.3=py310h055cbcc_0
numpy-base=1.26.3=py310h65a83cf_0
openjpeg=2.4.0=h4fc8c34_0
openssl=3.0.13=h2bbff1b_0
packaging=23.2=pypi_0
pandas=2.2.0=pypi_0
pillow=10.0.1=py310h045eedc_0
pip=23.3.1=py310haa95532_0
protobuf=4.25.2=pypi_0
pyarrow=15.0.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pycryptodome=3.20.0=pypi_0
pydantic=2.6.1=pypi_0
pydantic-core=2.16.2=pypi_0
pymilvus=2.3.6=pypi_0
pyopenssl=23.2.0=py310haa95532_0
pysocks=1.7.1=py310haa95532_0
python=3.10.13=he1021f5_0
python-dateutil=2.8.2=pypi_0
python-dotenv=1.0.1=pypi_0
pytorch=2.2.0=py3.10_cpu_0
pytorch-mutex=1.0=cpu
pytz=2024.1=pypi_0
pyyaml=6.0.1=py310h2bbff1b_0
regex=2023.12.25=pypi_0
requests=2.31.0=py310haa95532_0
s3transfer=0.10.0=pypi_0
setuptools=68.2.2=py310haa95532_0
six=1.16.0=pypi_0
sniffio=1.3.0=pypi_0
sqlite=3.41.2=h2bbff1b_0
starlette=0.36.3=pypi_0
sympy=1.12=py310haa95532_0
tbb=2021.8.0=h59b6b97_0
tk=8.6.12=h2bbff1b_0
torchaudio=2.2.0=py310_cpu
torchvision=0.17.0=py310_cpu
tqdm=4.66.1=pypi_0
typing_extensions=4.9.0=py310haa95532_1
tzdata=2023.4=pypi_0
ujson=5.9.0=pypi_0
urllib3=1.26.18=py310haa95532_0
uvicorn=0.27.0.post1=pypi_0
vc=14.2=h21ff451_1
vine=5.1.0=pypi_0
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.13=pypi_0
wheel=0.41.2=py310haa95532_0
win_inet_pton=1.1.0=py310haa95532_0
xz=5.4.5=h8cc25b3_0
yaml=0.2.5=he774522_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
